Version 1.02 from 2015/11/01
For almost every load-bearing structure, forces have to be induced or
diverted in a way, that the transmission of force is to be configured.
The engidesk App at hand allows for the pre-dimensioning of those loop
connections in which UD-layers wind around a bolt in form of a parallel
loop.